make sure cache_key exists

+++ modules.local/ppcp-wc-gateway/src/Settings/SettingsListener.php
@@ -44,7 +44,9 @@ class SettingsListener
             $this->settings->reset();
             $this->settings->persist();
             $this->webhookRegistrar->unregister();
-            $this->cache->delete(PayPalBearer::CACHE_KEY);
+            if ($this->cache->has(PayPalBearer::CACHE_KEY) ) {
+                $this->cache->delete(PayPalBearer::CACHE_KEY);
+            }
             return;
         }
 
@@ -60,7 +62,9 @@ class SettingsListener
             $this->settings->set($id, $value);
         }
         $this->settings->persist();
-        $this->cache->delete(PayPalBearer::CACHE_KEY);
+        if ($this->cache->has(PayPalBearer::CACHE_KEY) ) {
+            $this->cache->delete(PayPalBearer::CACHE_KEY);
+        }
     }
 
     //phpcs:disable Inpsyde.CodeQuality.NestingLevel.MaxExceeded
